refactor: ajustments e style classes
