+++ MyTheFourth/src/MyTheFourth.Frontend/DevsResistenceContext/Constants/MyTheFourthHttpServiceEndpoints.cs
@@ -1,0 +1,10 @@
+namespace MyTheFourth.Frontend.DevsResistenceContext.Constants;
+
+public static class MyTheFourthHttpServiceEndpoints
+{
+    public const string MoviesEndpoint = "/movies";
+    public const string CharacterEndpoint = "/characters";
+    public const string PlanetsEndpoint = "/planets";
+    public const string StarshipsEndpoint = "/starships";
+    public const string VehiclesEndpoint = "/starships";
+}
