+++ MyTheFourth/src/MyTheFourth.Frontend/DevsResistenceContext/Models/CharacterListResponse.cs
@@ -1,0 +1,6 @@
+namespace MyTheFourth.Frontend.DevsResistenceContext.Models;
+
+public class CharacterListResponse : ListResponse<CharacterDataModel>
+{
+
+}
