+++ MyTheFourth/src/MyTheFourth.Frontend/DevsResistenceContext/Models/MovieListResponse.cs
@@ -1,0 +1,6 @@
+namespace MyTheFourth.Frontend.DevsResistenceContext.Models;
+
+public class MovieListResponse : ListResponse<MovieDataModel>
+{
+
+}
