+++ MyTheFourth/src/MyTheFourth.Frontend/DevsResistenceContext/Models/PlanetListResponse.cs
@@ -1,0 +1,6 @@
+namespace MyTheFourth.Frontend.DevsResistenceContext.Models;
+
+public class PlanetListResponse : ListResponse<PlanetDataModel>
+{
+
+}
