+++ MyTheFourth/src/MyTheFourth.Frontend/DevsResistenceContext/Models/StarshipListResponse.cs
@@ -1,0 +1,6 @@
+namespace MyTheFourth.Frontend.DevsResistenceContext.Models;
+
+public class StarshipListResponse : ListResponse<StarshipDataModel>
+{
+
+}
