+++ MyTheFourth/src/MyTheFourth.Frontend/DevsResistenceContext/Models/VehicleListResponse.cs
@@ -1,0 +1,6 @@
+namespace MyTheFourth.Frontend.DevsResistenceContext.Models;
+
+public class VehicleListResponse : ListResponse<VehicleDataModel>
+{
+
+}
